create edit form component
