+++ src/scenes/Reports/AcceleratorReportEdit.tsx
@@ -1,29 +1,22 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
-import { Box, Typography, useTheme } from '@mui/material';
-
-import AcceleratorReportStatusBadge from 'src/components/AcceleratorReports/AcceleratorReportStatusBadge';
-import MetricBox, { getMetricId } from 'src/components/AcceleratorReports/MetricBox';
 import Page from 'src/components/Page';
-import Card from 'src/components/common/Card';
-import WrappedPageForm from 'src/components/common/PageForm';
 import TitleBar from 'src/components/common/TitleBar';
-import useNavigateTo from 'src/hooks/useNavigateTo';
 import { useLocalization } from 'src/providers';
 import { useParticipantData } from 'src/providers/Participant/ParticipantContext';
 import { selectListAcceleratorReports } from 'src/redux/features/reports/reportsSelectors';
 import { requestListAcceleratorReports } from 'src/redux/features/reports/reportsThunks';
 import { useAppDispatch, useAppSelector } from 'src/redux/store';
 import strings from 'src/strings';
-import { AcceleratorReport, MetricType } from 'src/types/AcceleratorReport';
+import { AcceleratorReport } from 'src/types/AcceleratorReport';
 
-const AcceleratorReportView = () => {
+import AcceleratorReportEditForm from './AcceleratorReportEditForm';
+
+const AcceleratorReportEditView = (): JSX.Element | null => {
   const { activeLocale } = useLocalization();
   const { currentParticipantProject, setCurrentParticipantProject } = useParticipantData();
-  const theme = useTheme();
   const dispatch = useAppDispatch();
-  const { goToAcceleratorReport } = useNavigateTo();
 
   const pathParams = useParams<{ projectId: string; reportId: string }>();
   const reportId = String(pathParams.reportId);
@@ -75,6 +68,10 @@ const AcceleratorReportView = () => {
   const reportName =
     selectedReport?.frequency === 'Annual' ? year : selectedReport?.quarter ? `${year}-${selectedReport?.quarter}` : '';
 
+  if (!selectedReport) {
+    return null;
+  }
+
   return (
     <Page
       title={
@@ -88,67 +85,9 @@ const AcceleratorReportView = () => {
         />
       }
     >
-      <WrappedPageForm
-        cancelID={'cancelEditAcceleratorReport'}
-        onCancel={() => {
-          goToAcceleratorReport(Number(reportId), Number(projectId));
-        }}
-        onSave={() => {
-          // TODO: save report
-        }}
-        saveID={'saveEditAcceleratorReport'}
-        saveDisabled={false}
-      >
-        <Box display='flex' flexDirection='column' flexGrow={1} overflow={'auto'}>
-          <Card
-            style={{
-              display: 'flex',
-              flexDirection: 'column',
-              flexGrow: 1,
-            }}
-          >
-            {selectedReport?.startDate && selectedReport?.endDate && (
-              <Box
-                borderBottom={`1px solid ${theme.palette.TwClrBrdrTertiary}`}
-                padding={theme.spacing(3, 0)}
-                marginBottom={3}
-              >
-                <div style={{ float: 'right', marginBottom: '0px', marginLeft: '16px' }}>
-                  <AcceleratorReportStatusBadge status={selectedReport.status} />
-                </div>
-
-                <Typography fontSize={14} fontStyle={'italic'}>
-                  {strings.formatString(strings.REPORT_PERIOD, selectedReport?.startDate, selectedReport?.endDate)}
-                </Typography>
-              </Box>
-            )}
-            {['system', 'project', 'standard'].map((type) => {
-              const metrics =
-                type === 'system'
-                  ? selectedReport?.systemMetrics
-                  : type === 'project'
-                    ? selectedReport?.projectMetrics
-                    : selectedReport?.standardMetrics;
-
-              return metrics?.map((metric, index) => (
-                <MetricBox
-                  editingId={getMetricId(metric, type as MetricType)}
-                  index={index}
-                  key={`${type}-${index}`}
-                  metric={metric}
-                  projectId={projectId}
-                  reload={reload}
-                  reportId={Number(reportId)}
-                  setEditingId={() => {}}
-                  type={type as MetricType}
-                />
-              ));
-            })}
-          </Card>
-        </Box>
-      </WrappedPageForm>
+      <AcceleratorReportEditForm report={selectedReport} />
     </Page>
   );
 };
 
-export default AcceleratorReportView;
+export default AcceleratorReportEditView;
